Basic, basic working ncurses version.
It ain't pretty, it ain't portable, it ain't bulletproof but you can
successfully have an irc session and make it work.

To make it pretty, I might have to make a detour into the mysterious
land of FFI or otherwise port over the Curses Development Kit. It may
not be particularly glamorous itself, but it'll make life a lot easier.